requirements.txt
requests==2.31.0
web3==6.15.0
py-clob-client==0.1.0
python-dotenv==1.0.0
